tests/wpt/tests/css/css-multicol/offsetProps-002.html

<!DOCTYPE html>
<title>offsetTop inside relpos block in inline in multicol</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://issues.chromium.org/issues/434678447">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div style="columns:1; margin-top:100px;">
  <div style="height:100px;"></div>
  <div style="display:inline;">
    <div id="expectedOffsetParent" style="position:relative; border:solid;">
      <div id="target" style="margin-top:10px; height:20px;"></div>
    </div>
  </div>
</div>
<script>
  test(() => {
    assert_equals(target.offsetParent, expectedOffsetParent);
    assert_equals(target.offsetTop, 10);
  }, "offsetTop");
</script>
